MTP-1738 Split up huge test file into two to allow CircleCI to better allocate tests to the parallelised test instances
